CLeanups
git-svn-id: svn+ssh://solsvn.oma.be/home/sidc/ssr/SWHV_repos/trunk/client/jhv-3d-wcs@5504 b4e469a2-07ce-4b26-9273-4d7d95a670c7

+++ src/plugins/swhvhekplugin-3d/src/org/helioviewer/jhv/plugins/swhvhekplugin/SWHVHEKPluginRenderable.java
@@ -75,7 +75,7 @@ public class SWHVHEKPluginRenderable implements Renderable {
         }
 
         double arcResolution = 100;
-        int lineResolution = 2;
+        double lineResolution = 2;
 
         Date date = new Date((evt.getStartDate().getTime() + evt.getEndDate().getTime()) / 2);
         Position.Latitudinal p = Sun.getEarth(date);
